Refactor play/pause observer to singleton to avoid multiple observers on models for the same player.
