+++ Pocket Casts TV App/UI/Player/NowPlayingViewModel.swift
@@ -68,10 +68,10 @@ class NowPlayingViewModel: Identifiable {
             return
         }
 
+        PlayerStatusAnalytics.shared.observe(player: player)
         timeControlStatusObservation = player.observe(\.timeControlStatus, options: [.new, .initial]) { [weak self] _, _ in
             Task { @MainActor [weak self] in
                 self?.updateLoadingState()
-                self?.updatePlayState()
             }
         }
 
@@ -92,46 +92,6 @@ class NowPlayingViewModel: Identifiable {
             Task { @MainActor [weak self] in
                 self?.updateLoadingState()
             }
-        }
-    }
-
-    private var previousTimeStatus: AVPlayer.TimeControlStatus?
-
-    private func updatePlayState() {
-        guard let player else {
-            return
-        }
-        let currentTimeStatus = player.timeControlStatus
-        if previousTimeStatus == nil {
-            //ignore first change on load, or changes that are not coming from play/pause
-            previousTimeStatus = currentTimeStatus
-            return
-        }
-
-        guard previousTimeStatus != currentTimeStatus, previousTimeStatus == .paused || previousTimeStatus == .playing else {
-            previousTimeStatus = currentTimeStatus
-            return
-        }
-        previousTimeStatus = currentTimeStatus
-
-        if let interval = AnalyticsPlaybackHelper.shared.timestampOfLastRemoteAction?.timeIntervalSinceNow,
-             abs(interval) < 1 {
-            // Do not need to track playback changes made by remote just now
-            AnalyticsPlaybackHelper.shared.timestampOfLastRemoteAction = nil
-            return
-        }
-
-        switch currentTimeStatus {
-        case .playing:
-            AnalyticsPlaybackHelper.shared.currentSource = .player
-            AnalyticsPlaybackHelper.shared.play()
-        case .paused:
-            AnalyticsPlaybackHelper.shared.currentSource = .player
-            AnalyticsPlaybackHelper.shared.pause()
-        case .waitingToPlayAtSpecifiedRate:
-            break
-        @unknown default:
-            break
         }
     }
